backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic==2.7.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
